Move CompileM to quasar.fs
