+++ connector/src/main/scala/quasar/connector/package.scala
@@ -16,15 +16,11 @@
 
 package quasar
 
-import quasar.common.PhaseResultW
 import quasar.effect.Failure
-import quasar.frontend.SemanticErrsT
 
 import scalaz._
 
 package object connector {
-  type CompileM[A] = SemanticErrsT[PhaseResultW, A]
-
   type EnvErr[A] = Failure[EnvironmentError, A]
   type EnvErrT[F[_], A] = EitherT[F, EnvironmentError, A]
